+++ src/Event/FieldOptionsEvent.php
@@ -9,14 +9,18 @@ use Symfony\Contracts\EventDispatcher\Event;
 class FieldOptionsEvent extends Event
 {
     private bool $dirty = false;
-
     private bool $emptyOption = false;
+    private ?bool $grouped = null;
+    private bool $autoGrouped = false;
+    private bool $sorted = false;
 
     private string $emptyOptionLabel = '-';
 
-    public function __construct(private readonly Widget $widget, private readonly Form $form, private array $options = [])
-    {
-    }
+    public function __construct(
+        private readonly Widget $widget,
+        private readonly Form   $form,
+        private array           $options = []
+    ) {}
 
     public function getWidget(): Widget
     {
@@ -28,28 +32,36 @@ class FieldOptionsEvent extends Event
         return $this->form;
     }
 
-    public function addOption(string $value, ?string $label = null): void
+    public function addOption(string $value, ?string $label = null, ?string $group = null): self
     {
-        $this->options[] = $this->createOptions($value, $label);
+        $this->options[] = $this->createOptions($value, $label, $group);
         $this->dirty = true;
+
+        if (!\is_null($group)) {
+            $this->autoGrouped = true;
+        }
+
+        return $this;
     }
 
-    public function createOptions(string $value, ?string $label = null): array
+    public function createOptions(string $value, ?string $label = null, ?string $group = null): array
     {
         if (!$label) {
             $label = $value;
         }
 
         return [
+            "group" => $group,
             "value" => $value,
             "label" => $label,
         ];
     }
 
-    public function setOptions(array $options = []): void
+    public function setOptions(array $options): self
     {
         $this->options = $options;
         $this->dirty = true;
+        return $this;
     }
 
     public function getOptions(): array
@@ -62,13 +74,36 @@ class FieldOptionsEvent extends Event
         return $this->dirty;
     }
 
+    public function isGrouped(): bool
+    {
+        return $this->grouped ?? $this->autoGrouped;
+    }
+
+    public function setGrouped(?bool $grouped): self
+    {
+        $this->grouped = $grouped;
+        return $this;
+    }
+
+    public function isSorted(): bool
+    {
+        return $this->sorted;
+    }
+
+    public function setSorted(bool $sorted): self
+    {
+        $this->sorted = $sorted;
+        return $this;
+    }
+
     /**
-     * Set options by values. If a reference array is given, the values will be used as keys and the reference array as values.
+     * Set options by values. If a reference array is given, the values will be used as keys and the reference array as
+     * values.
      *
      * @param array $values The option values
-     * @param array $reference A optional language array
+     * @param array $reference An optional language array
      */
-    public function setOptionsByValues(array $values, array $reference = []): void
+    public function setOptionsByValues(array $values, array $reference = []): self
     {
         $this->setOptions([]);
         foreach ($values as $option) {
@@ -76,9 +111,10 @@ class FieldOptionsEvent extends Event
         }
 
         $this->dirty = true;
+        return $this;
     }
 
-    public function setOptionsByKeyValue(array $options): void
+    public function setOptionsByKeyValue(array $options): self
     {
         $this->setOptions([]);
         foreach ($options as $key => $value) {
@@ -86,12 +122,14 @@ class FieldOptionsEvent extends Event
         }
 
         $this->dirty = true;
+        return $this;
     }
 
-    public function setEmptyOption(bool $emptyOption, string $label = '-'): void
+    public function setEmptyOption(bool $emptyOption, string $label = '-'): self
     {
         $this->emptyOption = $emptyOption;
         $this->emptyOptionLabel = $label;
+        return $this;
     }
 
     public function isEmptyOption(): bool